Split udev functions to dedicated module
This patch moves existing udev related functions to src/udev-support.[ch].
While doing this I realized I also need to move verify_appname() so for the
lack of a better place I moved it to src/snap.[ch].

I plan to move more functions out of main, to the point where each of the
support modules has an API for doing one thing it is responsible for and other
functions can be static.
