+++ src/Makefile.am
@@ -1,10 +1,19 @@
 bin_PROGRAMS = snap-confine
-snap_confine_SOURCES = main.c utils.c utils.h
+snap_confine_SOURCES = \
+	main.c \
+	utils.c \
+	utils.h \
+	snap.c \
+	snap.h
 snap_confine_CFLAGS = -Wall -Werror
 snap_confine_LDADD =
 
 if STRICT_CONFINEMENT
-snap_confine_SOURCES += seccomp_utils.c seccomp_utils.h
+snap_confine_SOURCES += \
+	seccomp_utils.c \
+	seccomp_utils.h \
+	udev-support.c \
+	udev-support.h
 snap_confine_CFLAGS += $(APPARMOR_CFLAGS) $(SECCOMP_CFLAGS) $(UDEV_CFLAGS)
 snap_confine_LDADD += $(APPARMOR_LIBS) $(SECCOMP_LIBS) $(UDEV_LIBS)
 endif
